+++ tests/config.lua
@@ -110,6 +110,10 @@ return {
           callback = "strategies.chat.tools.catalog.create_file",
           description = "Create a file in the current working directory",
         },
+        ["delete_file"] = {
+          callback = "strategies.chat.tools.catalog.delete_file",
+          description = "Delete a file in the current working directory",
+        },
         ["fetch_webpage"] = {
           callback = "strategies.chat.tools.catalog.fetch_webpage",
           description = "Fetches content from a webpage",
@@ -322,7 +326,7 @@ return {
       },
       slash_commands = {
         ["buffer"] = {
-          callback = "strategies.chat.slash_commands.buffer",
+          callback = "strategies.chat.slash_commands.catalog.buffer",
           description = "Insert open buffers",
           keymaps = {
             modes = {
@@ -336,7 +340,7 @@ return {
           },
         },
         ["fetch"] = {
-          callback = "strategies.chat.slash_commands.fetch",
+          callback = "strategies.chat.slash_commands.catalog.fetch",
           description = "Insert URL contents",
           opts = {
             adapter = "jina", -- jina|tavily
@@ -345,7 +349,7 @@ return {
           },
         },
         ["file"] = {
-          callback = "strategies.chat.slash_commands.file",
+          callback = "strategies.chat.slash_commands.catalog.file",
           description = "Insert a file",
           opts = {
             contains_code = true,